requirements.txt
attrs
aiohttp
htmlwrite
sockjs
pyyaml
python-slugify
uvloop
msgpack-python
